Transformar componente Minhas ACCs em funcional
